requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-appconf==1.0.4
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-imagekit==4.0.2
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-newsletter==0.9.1
django-tinymce==3.0.2
idna==2.10
ldif3==3.1.1
oauthlib==3.1.0
pilkit==2.0
python-card-me==0.9.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
surlex==0.2.0
unicodecsv==0.14.1
urllib3==1.25.10
whitenoise==5.1.0
